refactor: Use memory buffers instead of streams for results
- Result objects now only know and care about the result payload part of
  a result cache entry.
- Result object are no longer tightly coupled with the primary storage
  implementation.

This is part of a larger refactoring effort with the goal of simplifying
how cache entries are read and processed.

+++ src/core/CacheEntryHeader.hpp
@@ -76,7 +76,7 @@ struct CacheEntryHeader
   std::string namespace_;
   uint64_t entry_size;
 
-  uint64_t payload_size() const;
+  uint32_t payload_size() const;
   void set_entry_size_from_payload_size(uint64_t payload_size);
   void inspect(FILE* stream) const;
